LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/form-double-submit-to-different-origin-frame-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://web-platform.test:8800" from accessing a frame with origin "http://www1.web-platform.test:8800". Protocols, domains, and ports must match.
CONSOLE MESSAGE: Blocked a frame with origin "http://web-platform.test:8800" from accessing a frame with origin "http://www1.web-platform.test:8800". Protocols, domains, and ports must match.
This frame should stay blank

This frame should navigate (to 404)



PASS default submit action should supersede onclick submit() for cross-origin iframes
